server: Use correct atom name to register mxim with X server
XIM servers have to use an atom starting with the name "@server=" to
register with the X server, but mxim uses an atom starting with "@im=",
making clients unable to connect to mxim.

This commit changes the name of the atom that mxim uses to register
with the X server from "@im=mxim" to "@server=mxim".