server: Make mxim owner of XIM_SERVER and "@server=mxim" atoms
When a client connects to a XIM server, it looks at the owner of
the atom that the XIM_SERVERS atom points to, to determine whom
to talk to.

This commit modifies the server so that it creates a window and
makes it the owner of the XIM_SERVERS and "@server=mxim" atoms.