server: Use selections to tell clients the server address and locales
To determine the locales and address of the XIM server, clients will
send a XSelectionRequestEvent for the "LOCALES" and "TRANSPORT" atoms
to the XIM Server.

This commit modifies the server so that it handles X events of type
XSelectionRequestEvent, telling clients the supported locales and
protocol/address of the XIM server.