Add abstract socket implementation
A socket base type is necessary to facilitate the implementation of
an event loop that can handle server, client, and X11 connections in
the same way.

This commit adds an abstract socket base type that can be used for
the implementation of XIM server and client sockets, as well as X11
connections.