ximclient: Remove type to track client state
The state of client connections is tracked using a variable with two
possible values, which is no better than using a pointer to the input
method that is set when a XIM_OPEN message is received.

This commit removes the variable and type that are used to track the
state of client connections.