inputmethod: Add type to manage input methods
When a XIM_OPEN message is received, the XIM Server should look up
an adequate input method for the locale that is specified in that
message, and associate the input method with the client connection.

This commit adds an `input_method` type that will be used to handle
input methods and associate them with clients.