ximproto: Return correct message length in `xim_msg_decode()`
The message length returned by `xim_msg_decode()` does not include
the length of the header, causing `_xim_client_in()` to skip over
the wrong amount of data when determining the start of the next
message.

This commit modifies `xim_msg_decode()` so that it includes the
header when calculating the length of the parsed message.