ximclient: Implement handling of XIM_QUERY_EXTENSION messages
XIM clients cannot connect to mxim because it does not implement
handling of XIM_QUERY_EXTENSION messages.

This commit modifies the `xim_client` implementation so that it
handles XIM_QUERY_EXTENSION messages and replies with a list of
extensions that is supported by the input method that is associated
with the client.