inputmethod: Allow input methods to declare supported encodings
Mxim needs to know the input methods' supported encodings to agree
on an encoding with clients, but there is no way for input methods
to declare their supported encodings.

This commit adds a member to the `input_method` type that allows
input methods to declare their supported encodings as an array of
string pointers. The array shall be sorted in order of preference,
the first element being the most preferable encoding.