ximproto: Parse the byte order in XIM_CONNECT messages
Because `decode_XIM_CONNECT()` does not parse the byte order of the
received message, mxim cannot determine if it uses the same endianness
as the client.

This commit modifies `decode_XIM_CONNECT()` so that it sets the
`byte_order` field of the `xim_msg_connect_t` according to the
message received from the client.