inputmethod: Allow attribute values to be associated with input method
The `input_method` type does not allow values of attributes to be
associated with it, making it necessary that input methods handle
requests to get and set attributes themselves, instead of leaving
it to mxim.

This commit adds pointers to the `input_method` type so that input
methods can associate memory for attribute values with it, allowing
mxim to manipulate them directly.