ximclient: Handle XIM_ENCODING_NEGOTIATION messages
Clients cannot negotiate the encoding with mxim because it does not
handle XIM_ENCODING_NEGOTIATION messages.

This commit modifies mxim so that it handles XIM_ENCODING_NEGOTIATION
messages. When such a message is received, it compares the lists of
encodings supported by the input method and the client and sends a
XIM_ENCODING_NEGOTIATION_REPLY message with the mutually supported
encoding that is most favored by the input method.