ximclient: Send XIM_ERROR messages when errors occur
Mxim does not correctly signal errors to clients because it didn't
have a function to send XIM_ERROR messages. It now has a function
for sending error messages, so it should use it.

This commit modifies the `xim_client` implementation so that it
sends XIM_ERROR messages when errors occur.