ximproto,ximclient: Adapt to changes of input_method structure
The way attributes and attribute values are stored in `input_method`
structures has changed, and code that is interfacing with input_method
types needs to be updated.

This commit adapts the `xim_msg_open_reply_t` structure and handlers of
XIM_OPEN and XIM_GET_IM_VALUES messages to the new structure of the
`input_method` type.