ximclient: Set event mask when client opens an input method
Clients do not forward events to mxim because it does not set an
event mask when an IM is opened or an IC is created.

This commit modifies the XIM_OPEN handler so that it sends a
XIM_SET_EVENT_MASK message to the client, telling it to forward all
key presses to mxim.