inputmethod: Use correct types in signature of event handler
The signature of the `event` handler of `input_method_t` suggests
that event data is passed in an XEvent structure, but it is passed
in the wire format of the X Core Protocol.

This commit updates the signature of the `event` handler of
`input_method_t` so that it uses the correct types.