ximclient: Add handler for XIM_FORWARD_EVENT messages
Input methods cannot handle events from clients because mxim does not
handle XIM_FORWARD_EVENT messages.

This commit adds a handler for XIM_FORWARD_EVENT messages that passes
events to the client's input method's `event` handler and sends the
appropriate response to the client. If the handler requests the event
to be filtered, it sends only a XIM_SYNC_REPLY message to the client.
Otherwise, it unsets the synchronous flag of the XIM_FORWARD_EVENT
message and sends it back to the client, followed by a XIM_SYNC_REPLY
message.