ximclient: Pass keysym_t to event handler
Input methods' `event` handlers are passed key events in the X Core
Protocol's wire format, which makes it awkward to handle the events.

This commit modifies the `input_method_t` type and `xim_client`
implementation so that a `keysym_t` is passed to the handler,
instead of the event structure from the X Core Protocol.