inputmethod: Add event handler that is called when an IC is created
When an Input Context is created, Input Methods need to allocate
preedit buffers and perform other initialization, but there is no
way for the Input Method to get notified when an Input Context is
allocated.

This commit adds the event handler `ic_created` to the `input_method`
type, allowing input methods to perform IM-specific initialization
when an IC is created.