keysym: Fit modmask_t values into 4 bits
There are only 16 possible values for the modmask_t type, but the
mask with all modifiers occupies 6 bits. To implement key bindings
with a naive table thus uses four times the memory of a table with
dense modmasks.

This commit changes the values of modmask_t so that only four bits
are necessary to represent all possible combinations of modmasks.