inputcontext: Add reference to the corresponding XIM client
The `input_context_t` type does not contain a reference to the XIM
client that it belongs to, making it impossible to implement sending
XIM_COMMIT messages without separately tracking XIM clients and ICs.

This commit modifies the `input_context_t` type so that it contains
a reference to the XIM client that it belongs to.