inputmethod: Add function to handle key events
Handling of key events is not implemented by the `input_method_t`
type, making it necessary for input methods to handle keys directly.
However, key event handling is not specific to the individual input
methods, and so it should be handled by the `input_method_t` type.

This commit adds the `input_method_handle_key()` method, which is
to be invoked by `xim_client_t` when a key event is received from
the client.