ximclient: Pass client to input context constructor
The signature of `input_context_new()` has been changed so that the
caller has to pass a pointer to the `xim_client_t` instance that the
IC belongs to, but `handle_create_ic_msg()` still expects the IC
constructor to have the old signature.

This commit modifies the `handle_create_ic_msg()` function so that
it passes a client reference to `input_context_new()`.