ximclient: Add functions for retrieving IMs and ICs
Methods of the `input_method_t` and `input_context_t` types need to
be able to look up input methods and input contexts associated with
XIM clients, but there are no methods to do that.

This commit adds the `xim_client_get_im()` and `xim_client_get_ic()`
which allow the caller to look up IMs and ICs by their id.