ximclient: Implement method for sending XIM_SYNC messages
The `ximclient_t` type has a method for sending XIM_SYNC_REPLY messages,
but there is no equivalent method for XIM_SYNC messages.

This commit renames the `send_sync_reply()` method to `xim_client_sync()`
and changes it so that it can be used to send XIM_SYNC and XIM_SYNC_REPLY
messages.