ximclient: Fix synchronization in XIM_FORWARD_EVENT handling
Mxim does not correctly handle synchronization when it handles
XIM_FORWARD_EVENT messages, causing the XIM client to hang when
the IM commits a preedit buffer.

This commit modifies the XIM_FORWARD_EVENT message handler so that
it determines whether a XIM_SYNC_REPLY has to be sent from the state
of the `synchronous` bit in the event message, and correctly sets
the `synchronous` bit in outgoing event messages.