segment: Add type to handle parts of the preedit buffer
CJK input methods typically allow a user to insert characters into
a preedit buffer, where it is then converted before the IM commits
the input to the application. However, depending on the amount of
text in the buffer, it may not be possible to convert the entire
buffer at once, but it may contain segments (think multiple words)
that need to be converted individually.

This commit adds the `segment_t` type which represents one unit of
user input that can be converted as a whole.