preedit: Correctly advance cursor if preedit is empty
The `_preedit_update_cursor()` function does not correctly advance
the cursor if the preedit buffer is empty, causing the second call
to `preedit_insert()` to overwrite the character written by the
first call.

This commit fixes the bounds check in `_preedit_update_cursor()`
so that it places the cursor after the last byte in the segment if
it is behind the segment.