keysym: Add modifier mask `MOD_NONE` to make code more readable
When defining key bindings, `MOD_SHIFT` and similar definitions can
be used to mean that a modifier has been pressed, but there is no
definition to mean "no modifier pressed", making it necessary to use
the more ambiguous `0`.

This commit adds the definition `MOD_NONE` which can be used as
modifier mask indicating that no modifier was pressed.