inputcontext: Use `preedit_t` to manage preedit buffer
The `input_context_t` type does not manage the preedit buffer, making
it necessary for the input method to allocate the preedit buffer and
to handle key inputs.

This commit modifies the `input_context_t` type so that it uses a
`preedit_t` to manage the preedit buffer. This commit further adds
the `input_context_insert()` and `input_context_erase()` methods
that allow characters to be added to and removed from the preedit
buffer.