inputcontext: Make input contexts language-aware
The main purpose of mxim is to allow the user multilingual input,
thus it needs to keep track of the input language either globally,
per client, or per input context. However, mxim is not language-
aware at all.

This commit modifies mxim so that it allows each input context to
have its own input language. This is necessary because mxim allows
XIM clients to have multiple active input contexts at a time, which
makes it unfeasible to switch languages globally or per-client.