inputmethod: Convert keysyms and insert them to the input context
The Input Method layer does not correctly pass on user inputs to the
input context, preventing user input from getting handled correctly.

This commit modifies the `input_method_handle_key()` function so that
it converts keysyms to the respective character in the IC's current
keymap and inserts the character into the input context's preedit
buffer.