config: Do not convert keysyms with modifiers other than shift
When looking up the character that corresponds to a keysym, the
`config_keysym_to_char()` function masks modifiers other than shift,
causing keysyms with control, alt, and super to be converted to a
character, while they should be ignored.

This commit modifies the `config_keysym_to_char()` function so it
does not convert keysyms with a modifier other than shift or no
modifier at all.