preedit: Merge `_preedit_update_cursor()` into `preedit_move()`
The `_preedit_update_cursor()` function was meant as a helper
function for `preedit_move()`, but it does all the work, and the
latter does nothing but call the former. Thus, the additional
call adds nothing but overhead.

This commit moves all code from `_preedit_update_cursor()` to
`preedit_move()` to remove the unnecessary call overhead.