preedit: Indicate if `preedit_move()` tried to move cursor beyond bounds
The `preedit_move()` function returns success, even if the caller
attempted to move the cursor beyond the bounds of the preedit buffer.
Thus, if the caller needs to detect this situation, they have to do so
manually.

This commit modifies the `preedit_move()` method so it returns `-ERANGE`
if the call would have resulted in the cursor getting moved beyond the
bounds of the preedit buffer.