char: Add function for conversion from `char_t` to UTF-8
Because mxim internally uses its own charset, user inputs need to
be converted to UTF-8 before they can be presented to the user or
sent to other applications. However, conversion from `char_t` to
UTF-8 is not implemented.

This commit adds the `char_to_utf8()` function which converts a
string of `char_t` characters to UTF-8.