preedit: Add method for retrieving user input
To provide visual feedback of the user input, a method to access the
user input that is stored in the preedit buffer is necessary.

This commit adds the `preedit_get_input()` method, which converts the
user input that is stored in the preedit to UTF-8 and writes it to a
caller-provided buffer.