string: Add type to make string manipulations more convenient
Displaying the state of the preedit buffer involves several string
manipulations, which is very inconvenient in C. A string type is
needed to make displaying the preedit buffer easier and less
error-prone.

This commit adds the `string_t` type, which can be used to easily
concatenate strings and replace substrings.