ximproto: Set number of values when decoding XIM_CREATE_IC message
When decoding XIM_CREATE_IC messages, the decoder does not set the
number of decoded XICATTRIBUTE objects, causing the message handler
to initialize an Input Context without attributes.

This commit modifies the `decode_XIM_CREATE_IC()` function so that
it counts the number of decoded attributes and sets the `num_values`
field of the decoded message appropriately.