xhandler: Add method for setting text properties on windows
The mwm window manager allows the MWM_HINT property of windows to
be used to display text in the indicator bar besides the window.
To display the state of the preedit buffer in the client window's
indicator bar, a function to set text properties of windows is
needed.

This commit adds the following two methods:
 - `x_handler_get_client_window()`: Returns the top-level window of
   a window (i.e. the window that the MWM_HINT property needs to be
   set on).
 - `x_handler_set_text_property()`: Sets the value of a window's
   property, encoded in `UTF8_STRING` format.