main: Make x_handler global
Window properties need to be set through an `x_handler_t` instance,
but there is only one such instance and it cannot be accessed from
outside of `main()`.

This commit makes the `x_handler_t` instance created in `main()`
global so that it can be used to set window properties from other
parts of mxim.