preedit: Add method for displaying preedit buffer contents to user
To display the contents of the preedit buffer to the user, a method
that converts the buffer contents to a UTF-8 encoded string is need.

This commit adds the `preedit_get_input_decorated()` method which
generates a representation of the preedit buffer with pango markup
that can be set as the top-level window's `MWM_HINT`.