preedit: Add method for retrieving the data to be committed
When committing the user input to the XIM client, mxim needs to query
the data to be committed from the preedit buffer, but there is no
method to do that.

This commit adds the `preedit_get_output()` method, which returns the
data to be committed (including autocomplete/conversion candidate
selection) from the preedit buffer.