inputmethod: Redraw input context if key was handled successfully
The `input_method_handle_key()` method does not redraw the input
context when a key was inserted. Thus, there is no visual feedback
until a command is executed.

This commit modifies the `input_method_handle_key()` method so that
it always redraws the input context if the key was successfully
handled.