segment: Implement clearing of candidates
The candidate list of a segment needs to be cleared when there are
no more candidates for a segment, but there is no method to clear
the list.

This commit modifies the `segment_set_candidates()` method so that
NULL can be passed in the `candidates` argument to clear the
segment's candidate list.