char: Add method for conversion from UTF-8 to `char_t`
To parse keys in dictionary files, a function to convert UTF-8 to
`char_t` is needed, but there is no such function.

This commit adds the `char_from_utf8()` method, which converts a
UTF-8 string to a `char_t` string. Character conversion uses a
lookup tree which is initialized when the function is called for
the first time.