parray: Allow multiple items to be passed to `parray_insert()`
The `parray_insert()` method reallocates the internal array every time
it is called. Because it can only insert one item at a time, this causes
significant overhead when inserting multiple items.

This commit modifies the `parray_insert()` method so that it can be used
to insert multiple items at a time.