segment: Use new dict types for displaying candidates
Segments expect candidates to be pointers to UTF-8 strings, but
candidates now have a distinct type that should be used instead.

This commit modifies the `segment_t` implementation so that the
`segment_set_candidates()` and `segment_get_candidates()` methods
expect `dict_candidate_t`-typed candidates.
Further, this adds the `segment_update_candidates()` method for
updating the candidate list when the contents of the segment have
changed.