inputcontext: Update candidates when the preedit buffer was changed
The autocompletion/conversion candidates need to be updated when the
preedit buffer of an input context was modified, but the method for
updating the candidates is never called.

This commit modifies the `input_context_t` implementation so that it
triggers an update of the candidates when a character was inserted
into or erased from the preedit buffer.