segment: Attempt to combine characters when inserting
In Japanese, some kana are created by adding dakuten or handakuten
to the previous characters. Thus, when inserting a character, mxim
should attempt to combine it with the previous character.

This commit modifies the `segment_insert()` method so that it
attempts to combine the new character with the character under the
cursor.