preedit: Don't move the cursor if characters were combined during insert
The `preedit_insert()` method attempts to move the cursor, even if the
insertion operation did not cause the length of the preedit buffer to
change, for example because characters were combined.

This commit modifies the `preedit_insert()` method so that it does not
call `preedit_move()` if the `segment_insert()` call did not add any
characters to the preedit buffer.