token: Buffer reads to improve lexer performance
The lexer reads dictionaries one byte per `read()` call, slowing it
down significantly. Buffering should be used to reduce the number of
syscalls made to parse dictionary files.

This commit modifies the lexer so that it reads the contents of the
dictionary to a buffer, from which bytes are passed to the caller.
This reduces the time it takes to parse a 17MB dictionary from 13s
to 3s on an Intel Core i5-2540M.