char: Do not use UTF-8 sequences directly as format strings
When converting `char_t` strings to UTF-8, UTF-8 sequences are used
directly as format strings in calls to `snprintf()`. This causes
incorrect results when the input contains single `%` characters as
`snprintf()` interprets them as incomplete conversion specifications.

This commit modifies the `char_to_utf8()` and `_estimate_size()`
functions so they do not pass UTF-8 sequences as format strings to
`snprintf()`.