inputcontext: Do not try to manipulate an empty preedit buffer
The `input_context_cursor_move()` and `input_context_erase()` methods
do not check if the preedit buffer is empty, leaving it up to the
`preedit_move()` and `preedit_erase()` functions how to behave and
what status to return to the caller.
Because mxim assumes that the request was not handled if an error was
returned, it forwards the unhandled key event to the client, which
causes a confusing user experience.

This commit modifies the `input_context_cursor_move()` and
`input_context_erase()` methods so that they return an error if the
input context's preedit buffer is empty.