preedit: Do not treat moves beyond preedit boundaries as errors
The `preedit_move()` method treats attempts to move beyond the
bounds of the preedit buffer as error. Because this method's return
value is propagated to upper layers and affects the decision whether
to forward or filter events from/to XIM clients, it affects the
user experience.

This commit modifies the `preedit_move()` method so that it does
not treat attempts to move beyond the preedit buffer's bounds as
an error. Instead, it will move the cursor the the start or end of
the preedit buffer and return success.