inputcontext: Do not move segment if preedit buffer is empty
The `input_context_move_segment()` method attempts to change the
current segment even if the preedit buffer is empty, causing the
key event that triggered the call to be filtered, even though it
should be forwarded to the XIM client.

This commit modifies the `input_context_move_segment()` method so
that it does not attempt to change the current segment if the
preedit buffer is empty.