Add command to activate and deactivate an input method
Because inputs are displayed in the window manager's pre-edit area,
users need a way to temporarily disable mxim, for example when typing
sensitive data such as passwords.

This commit gives `input_method` types an `active` field and adds the
CMD_ONOFF command, which allows the current input method's active
state to be toggled.