preedit: Automatically insert segments when inserting characters
Mxim does not support automatic insertion of segments, requiring the
user to insert segments manually, making the pre-editing needlessly
complicated.

This commit modifies the `preedit_insert()` method so that it inserts
a new segment for the new character if the new character and the one
that precedes it are not in the same character set.