cmd: Allow two shorts to be passed to a command
To implement jumping to the first or last segment of the preedit
buffer it is necessary to pass two shorts to a command, but the
`cmd_arg_t` union cannot store two shorts.

This commit adds a field to the `cmd_arg_t` union, allowing two
shorts to be assigned to it.