inputcontext: Allow segment movement in `input_context_cursor_move()`
Because the `input_context_cursor_move()` only moves the cursor
within a segment, but not between segments, it cannot be used to
implement jumping to the start or end of the preedit buffer.

This commit modifies the `input_context_cursor_move()` method so
that it can be used to move the cursor to another segment.