jkim: Pass correct arguments when invoking CMD_CURSOR_MOVE command
The `_jkim_cursor_move()` function does not pass the correct number
and types of arguments to the `input_context_cursor_move()` method
because the signature of the latter was changed.

This commit modifies the `_jkim_cursor_move()` function so that it
passes the correct arguments to `input_context_cursor_move()`.