config: Use CMD_CURSOR_MOVE command to move between segments
Mxim uses two separate commands to move the cursor inside a segment
and between segments, even though both can be implemented with the
same command.

This commit changes the configuration so that the `CMD_CURSOR_MOVE`
command is used to move the cursor inside a segment as well as
between segments.