preedit: Set the cursor offset when moving into another segment
When the cursor is moved into another segment, the offset into the
segment is not correctly adjusted, causing the cursor to jump.

This commit modifies the `preedit_move()` method so that it moves
the cursor to the right end of a segment when entering it from the
right, and to the left end of a segment when entering it from the
left.