inputcontext: Do not commit the preedit buffer if it is empty
The `input_context_commit()` method attempts to commit the preedit
buffer even if it is empty, causing it not to be ignored as it
ought to be. This causes presses of the return key not to be
forwarded to the X client when the preedit buffer is empty.

This commit modifies the `input_context_commit()` method so that it
returns an error if the preedit buffer is empty, causing mxim to
forward the triggering key press to the X client.