segment: Display selected candidate instead of input
The input text of a segment is displayed even when a candidate has
been selected for a segment. However, the candidate selection should
be shown to make it clear what is going to be committed.

This commit modifies the `segment_get_input_decorated()` method so
that it displays the selected candidate for a segment instead of the
user input, if the segment is not currently being edited.