hangul: Add functions to convert Hangul from char_t to UTF-8
Hangul are encoded with multiple characters when stored as char_t.
Thus, when converting hangul from char_t to UTF-8, the conversion
needs to combine multiple characters, but this is currently not
implemented.

This commit adds the `hangul_to_utf8()` method, which converts a
hangul from a char_t string to UTF-8.