preedit: Remove segments that were emptied by `preedit_erase()`
The `preedit_erase()` method does not remove empty segments, causing
the preedit to think it is not empty when it is. This causes mxim to
handle key strokes that should be forwarded to the X client instead
of handling them in the IM server.

This commit adds the `preedit_remove_segment()` method and modifies
`preedit_erase()` so that it removes the current segment if it was
emptied.