ximproto: Don't crash when handling XIM_DISCONNECT messages
The `xim_msg_decode()` does not allocate a buffer for XIM_DISCONNECT
messages because they don't contain a payload. However, the function
later attempts to dereference the message, causing it to dereference
a NULL pointer.

This commit modifies the `xim_decode_msg()` functions so that it
also allocates buffers for XIM_DISCONNECT messages.