hangul: Fix handling of compound jamo in patchim
When `hangul_to_utf8()` parses the patchim and finds two jamo that can
be stuffed into the patchim, it does so, even if the second jamo is
followed by a vowel, in which case it should be the start of a new
hangul instead.

This commit modifies the `hangul_to_utf8()` function so that it only
stuffs a second jamo into the patchim if it is not the start of a new
hangul.