xhandler: Don't terminate when X errors occur
When sending the contents of the preedit buffer to some X clients,
the X Server responds with errors, causing mxim to terminate. This
makes it hard to debug the source of the error.

This commit modifies the code that interacts with the X Server so
that mxim does not terminate when an error is received, but instead
prints an error message to stderr.