dict: Enable `dict_lookup()` to perform exact lookups
When mxim is going to conjugate results returned by `dict_lookup()`,
only exact matches may be used, but `dict_lookup()` also returns
matches that are longer than the user input.

This commit modifies `dict_lookup()` so that the caller has to
specify whether they want the lookup to be exact or predictive.