array: Move array functions to separate source unit
Functions for pointer array handling in `aide.c` may be useful for
other parts of mxim, so they should be moved into a separate source
unit.

This commit moves the `_array_add()` and `_array_free()` functions
to `array.c` so that they can be used by all parts of mxim.