array: Add methods for iterating over arrays and counting elements
There are no methods for counting or iterating over elements of an
array, making it necessary that code that handles arrays implements
such functions itself.

This commit adds the `array_foreach()` and `array_len()` methods,
which can be used to iterate over and count the number of elements
in an array, respectively.