aide,segment: Use suggestion_t type to handle input suggestions
Aide and segment methods use references to dictionary entries to
handle input suggestions, which makes it difficult to implement
conjugation of input suggestions.

This commit moddifies the aide and segment implementations so that
they use the new `suggestion_t` type to handle suggestions in a
mutable way, allowing mxim to modify suggestions before presenting
them to the user.