array: Allow caller of `array_free()` to specify element deallocator
The `array_free()` method frees elements contained within the array
with `free()`, which makes it impossible to use the method if elements
need to be freed with a different deallocator, or if they must not be
freed at all.

This commit modifies the `array_free()` method so that the caller
can specify the deallocator to be used to free array elements.