suggestion: Add priority and pointer to user data
Suggestions need to be ordered by relevance before presenting them
to the user, but the `suggestion_t` type does not contain data that
can be used for ordering.

This commit adds the `priority` and `data` fields to `suggestion_t`,
allowing suggestions to be sorted (for example using a parray) and
to store a reference to the dictionary candidate that the suggestion
was created from.