char: Correctly duplicate strings when `sizeof(char_t)` is not 1
When `char_dup()` copies the contents of the source string to the
duplicate string, it implicitly assumes that the size of char_t is
1, which is correct but not guaranteed never to change.

This commit modifies the `char_dup()` method so that it does not
implicitly assume `sizeof(char_t)` to be 1. This way, the method
will work correctly even when the size of char_t changes.