conjugation: Add type for storing words in their unconjugated form
Mxim will often process conjugated words as input. In these cases,
it has to figure out the dictionary form of the input and look up
the dictionary form instead of the user input. A data type is needed
to store the dictionary form and conjugate it back into the form that
was entered by the user.

This commit adds the `conjugation_t` type. This type is used to store
the dictionary form of a word as well as the suffix length and the
original conjugation, allowing the dictionary form to be conjugated
into the form entered by the user. This implementation can be used
for Japanese and Korean verbs and adjectives.