aide: Add support for conjugations of inputs
When handling Japanese and Korean, the user inputs will often be
conjugated words that cannot be directly looked up in the dictionary.
In those cases, mxim needs to figure out the dictionary form of the
user input and look up all possible candidates.

This commit modifies `aide_suggest()` and related functions so that
user inputs are first unconjugated, unconjugated words looked up in
the dictionary, and candidates conjugated back into the original form.
Conversion from user inputs into dictionary forms is not implemented
yet, and instead a null-conjugation is used that is equivalent with
a direct dictionary lookup as it was done until now.