conjugation: Properly terminate string in `conjugation_conjugate()`
The `conjugation_conjugate()` does not pass the proper buffer length
to the method that appends the conjugated suffix to the buffer, causing
it not to be properly terminated.

This commit modifies the `conjugation_conjugate()` method so that it
passes the correct buffer length to `char_to_utf8()`, causing the
buffer to be properly terminated with a 0-byte.