japanese: Set the conjugation type of i-adjectives
When the `get_i_adjective_conjugation()` method finds an i-adjective,
it does not set the type of the conjugation, rendering mxim unable to
use the conjugation to conjugate suggestions.

This commit modifies the `get_i_adjective_conjugation()` method so
that it sets the `type` member of any conjugations that it finds.