char: Allow char_concat() to be used on empty strings
The `char_concat()` method requires both strings to be non-empty,
making it necessary for the caller to perform such checks before
calling the method, needlessly complicating the caller's code.

This commit modifies the `char_concat()` method so that it can be
used on empty strings, making it behave like `char_dup()` if one
of the two strings to be concatenated is empty.
If two empty strings are passed to the method, it will return a
pointer to an empty string.