aide: Continue with next conjugation if a lookup didn't have matches
The `_lookup_conjugation()` method returns the return value of the
`array_foreach()` call that iterates over the results of the dictionary
lookup. If the lookup did not yield any results, an error is returned to
the calling `array_foreach()` invocation, which will terminate early
without iterating over the remaining conjugations, causing suggestions
to be incomplete.

This commit modifies the `_lookup_conjugation()` method so that it does
not treat a return value of `-ENOENT` from `array_foreach()` as error.