aide: Fix check if lookups should be predictive or exact
Mxim does not correctly check if the user input was determined to be
conjugated, causing it to make predictive lookups when exact lookups
are needed.

This commit modifies the `_lookup_conjugation()` method so that
the user input is determined to be conjugated if the conjugation
has a length bigger than zero.